# swagger.yml
basePath: /
consumes:
    - application/json
definitions:
...
    AuthParams:
        description: AuthParams - struct for auth params
        properties:
            id:
                type: string
                x-go-name: ID
            macaddress:
                type: string
                x-go-name: MacAddress
            password:
                type: string
                x-go-name: Password
        type: object
        x-go-package: github.com/gravitl/netmaker/models
...
    Iface:
        description: Iface struct for local interfaces of a node
        properties:
            address:
                $ref: '#/definitions/IPNet'
            addressString:
                type: string
                x-go-name: AddressString
            name:
                type: string
                x-go-name: Name
        type: object
        x-go-package: github.com/gravitl/netmaker/models
...
    LegacyNode:
        description: LegacyNode - legacy struct for node model
        properties:
            accesskey:
                type: string
                x-go-name: AccessKey
            action:
                type: string
                x-go-name: Action
            address:
                type: string
                x-go-name: Address
            address6:
                type: string
                x-go-name: Address6
            allowedips:
                items:
                    type: string
                type: array
                x-go-name: AllowedIPs
            connected:
                type: string
                x-go-name: Connected
            defaultacl:
                description: == PRO ==
                type: string
                x-go-name: DefaultACL
            dnson:
                type: string
                x-go-name: DNSOn
            egressgatewaynatenabled:
                type: string
                x-go-name: EgressGatewayNatEnabled
            egressgatewayranges:
                items:
                    type: string
                type: array
                x-go-name: EgressGatewayRanges
            egressgatewayrequest:
                $ref: '#/definitions/EgressGatewayRequest'
            endpoint:
                type: string
                x-go-name: Endpoint
            expdatetime:
                format: int64
                type: integer
                x-go-name: ExpirationDateTime
            failover:
                type: string
                x-go-name: Failover
            failovernode:
                type: string
                x-go-name: FailoverNode
            firewallinuse:
                type: string
                x-go-name: FirewallInUse
            id:
                type: string
                x-go-name: ID
            ingressgatewayrange:
                type: string
                x-go-name: IngressGatewayRange
            ingressgatewayrange6:
                type: string
                x-go-name: IngressGatewayRange6
            interface:
                type: string
                x-go-name: Interface
            interfaces:
                items:
                    $ref: '#/definitions/Iface'
                type: array
                x-go-name: Interfaces
            internetgateway:
                type: string
                x-go-name: InternetGateway
            ipforwarding:
                type: string
                x-go-name: IPForwarding
            isdocker:
                type: string
                x-go-name: IsDocker
            isegressgateway:
                type: string
                x-go-name: IsEgressGateway
            ishub:
                type: string
                x-go-name: IsHub
            isingressgateway:
                type: string
                x-go-name: IsIngressGateway
            isk8s:
                type: string
                x-go-name: IsK8S
            ispending:
                type: string
                x-go-name: IsPending
            isrelay:
                type: string
                x-go-name: IsRelay
            isrelayed:
                type: string
                x-go-name: IsRelayed
            isserver:
                type: string
                x-go-name: IsServer
            isstatic:
                description: IsStatic - refers to if the Endpoint is set manually or dynamically
                type: string
                x-go-name: IsStatic
            lastcheckin:
                format: int64
                type: integer
                x-go-name: LastCheckIn
            lastmodified:
                format: int64
                type: integer
                x-go-name: LastModified
            lastpeerupdate:
                format: int64
                type: integer
                x-go-name: LastPeerUpdate
            listenport:
                format: int32
                type: integer
                x-go-name: ListenPort
            localaddress:
                type: string
                x-go-name: LocalAddress
            locallistenport:
                format: int32
                type: integer
                x-go-name: LocalListenPort
            macaddress:
                type: string
                x-go-name: MacAddress
            mtu:
                format: int32
                type: integer
                x-go-name: MTU
            name:
                type: string
                x-go-name: Name
            network:
                type: string
                x-go-name: Network
            networksettings:
                $ref: '#/definitions/Network'
            os:
                type: string
                x-go-name: OS
            ownerid:
                type: string
                x-go-name: OwnerID
            password:
                type: string
                x-go-name: Password
            persistentkeepalive:
                format: int32
                type: integer
                x-go-name: PersistentKeepalive
            publickey:
                type: string
                x-go-name: PublicKey
            relayaddrs:
                items:
                    type: string
                type: array
                x-go-name: RelayAddrs
            server:
                type: string
                x-go-name: Server
            traffickeys:
                $ref: '#/definitions/TrafficKeys'
            udpholepunch:
                type: string
                x-go-name: UDPHolePunch
            version:
                type: string
                x-go-name: Version
        type: object
        x-go-package: github.com/gravitl/netmaker/models
...
host: api.demo.netmaker.io
info:
    description: |-
        # API Usage

        Most actions that can be performed via API can be performed via UI. We recommend managing your networks using the official netmaker-ui project. However, Netmaker can also be run without the UI, and all functions can be achieved via API calls. If your use case requires using Netmaker without the UI or you need to do some troubleshooting/advanced configuration, using the API directly may help.

        # Authentication

        API calls must be authenticated via a header of the format -H “Authorization: Bearer <YOUR_SECRET_KEY>” There are two methods to obtain YOUR_SECRET_KEY: 1. Using the masterkey. By default, this value is “secret key,” but you should change this on your instance and keep it secure. This value can be set via env var at startup or in a config file (config/environments/< env >.yaml). See the [Netmaker](https://docs.netmaker.org/index.html) documentation for more details. 2. Using a JWT received for a node. This can be retrieved by calling the /api/nodes/<network>/authenticate endpoint, as documented below.
    title: Netmaker
    version: 0.21.1
paths:
...
    /api/hosts/{hostid}:
        delete:
            operationId: deleteHost
            parameters:
                - description: HostID
                  in: path
                  name: hostid
                  required: true
                  type: string
                  x-go-name: HostID
            responses:
                "200":
                    $ref: '#/responses/apiHostResponse'
            schemes:
                - https
            summary: Deletes a Netclient host from Netmaker server.
            tags:
                - hosts
        put:
            operationId: updateHost
            parameters:
                - description: HostID
                  in: path
                  name: hostid
                  required: true
                  type: string
                  x-go-name: HostID
            responses:
                "200":
                    $ref: '#/responses/apiHostResponse'
            schemes:
                - https
            summary: Updates a Netclient host on Netmaker server.
            tags:
                - hosts
    /api/hosts/{hostid}/networks/{network}:
        delete:
            operationId: deleteHostFromNetwork
            parameters:
                - description: hostid to add or delete from network
                  in: path
                  name: hostid
                  required: true
                  type: string
                  x-go-name: HostID
                - description: network
                  in: path
                  name: network
                  required: true
                  type: string
                  x-go-name: Network
            responses:
                "200":
                    $ref: '#/responses/okResponse'
            schemes:
                - https
            summary: Given a network, a host is removed from the network.
            tags:
                - hosts
        post:
            operationId: addHostToNetwork
            parameters:
                - description: hostid to add or delete from network
                  in: path
                  name: hostid
                  required: true
                  type: string
                  x-go-name: HostID
                - description: network
                  in: path
                  name: network
                  required: true
                  type: string
                  x-go-name: Network
            responses:
                "200":
                    $ref: '#/responses/okResponse'
            schemes:
                - https
            summary: Given a network, a host is added to the network.
            tags:
                - hosts
    /api/hosts/{hostid}/signalpeer:
        post:
            operationId: signalPeer
            parameters:
                - description: HostID
                  in: path
                  name: hostid
                  required: true
                  type: string
                  x-go-name: HostID
            responses:
                "200":
                    $ref: '#/responses/signal'
            schemes:
                - https
            summary: send signal to peer.
            tags:
                - hosts
    /api/hosts/{hostid}/sync:
        post:
            operationId: synchost
            parameters:
                - description: HostID
                  in: path
                  name: hostid
                  required: true
                  type: string
                  x-go-name: HostID
            responses:
                "200":
                    $ref: '#/responses/networkBodyResponse'
            schemes:
                - https
            summary: Requests a host to pull.
            tags:
                - hosts
    /api/hosts/{hostid}keys:
        post:
            operationId: updateKeys
            parameters:
                - description: HostID
                  in: path
                  name: hostid
                  required: true
                  type: string
                  x-go-name: HostID
            responses:
                "200":
                    $ref: '#/responses/networkBodyResponse'
            schemes:
                - https
            summary: Update keys for a network.
            tags:
                - hosts
...
    /api/nodes/{network}/{nodeid}:
        delete:
            operationId: deleteNode
            parameters:
                - in: path
                  name: network
                  required: true
                  type: string
                  x-go-name: Network
                - in: path
                  name: nodeid
                  required: true
                  type: string
                  x-go-name: NodeID
                - description: Node
                  in: body
                  name: node
                  schema:
                    $ref: '#/definitions/LegacyNode'
                  x-go-name: Node
            responses:
                "200":
                    $ref: '#/responses/nodeResponse'
            schemes:
                - https
            summary: Delete an individual node.
            tags:
                - nodes
        get:
            operationId: getNode
            parameters:
                - in: path
                  name: network
                  required: true
                  type: string
                  x-go-name: Network
                - in: path
                  name: nodeid
                  required: true
                  type: string
                  x-go-name: NodeID
            responses:
                "200":
                    $ref: '#/responses/nodeResponse'
            schemes:
                - https
            summary: Get an individual node.
            tags:
                - nodes
        put:
            operationId: updateNode
            parameters:
                - in: path
                  name: network
                  required: true
                  type: string
                  x-go-name: Network
                - in: path
                  name: nodeid
                  required: true
                  type: string
                  x-go-name: NodeID
                - description: Node
                  in: body
                  name: node
                  schema:
                    $ref: '#/definitions/LegacyNode'
                  x-go-name: Node
            responses:
                "200":
                    $ref: '#/responses/nodeResponse'
            schemes:
                - https
            summary: Update an individual node.
            tags:
                - nodes
...
    /api/nodes/adm/{network}/authenticate:
        post:
            operationId: authenticate
            parameters:
                - description: network
                  in: path
                  name: network
                  required: true
                  type: string
                  x-go-name: Network
                - description: AuthParams
                  in: body
                  name: auth_params
                  schema:
                    $ref: '#/definitions/AuthParams'
                  x-go-name: AuthParams
            responses:
                "200":
                    $ref: '#/responses/successResponse'
            schemes:
                - https
            summary: Authenticate to make further API calls related to a network.
            tags:
                - authenticate
...
produces:
    - application/json
responses:
...
    apiHostResponse:
        description: ""
        schema:
            $ref: '#/definitions/ApiHost'
...
    networkBodyResponse:
        description: ""
        schema:
            $ref: '#/definitions/Network'
    nodeResponse:
        description: ""
        schema:
            $ref: '#/definitions/LegacyNode'
...
    okResponse:
        description: ""
...
    signal:
        description: ""
        schema:
            $ref: '#/definitions/Signal'
...
    successResponse:
        description: ""
        schema:
            $ref: '#/definitions/SuccessResponse'
...
schemes:
    - https
swagger: "2.0"